+++ Modern JavaScript From The Beginning/Section 2: JavaScript Language Fundamentals/6. Using The Console/app.js
@@ -1,0 +1,36 @@
+// Logging to console
+console.log('It\'s a log command...');
+console.log(2020);
+console.log(true);
+
+var greeting = 'Hello JS';
+console.log(greeting);
+
+console.log([2, 0, 1, 3]);
+
+console.log({
+    a: 1,
+    b: 2
+});
+
+console.table({
+    a: 1,
+    b: 2
+});
+
+console.error('This is an error message');
+
+console.clear();
+
+console.warn('This is a warning message');
+
+console.time('Hey');
+
+console.log('It\'s a log command...');
+console.log('It\'s a log command...');
+console.log('It\'s a log command...');
+console.log('It\'s a log command...');
+console.log('It\'s a log command...');
+console.log('It\'s a log command...');
+
+console.timeEnd('Hey');